Separate the labeling and commenting logic
Addresses issue #10.

Prior to this commit, the bot was interacting with GH issues on a Gerrit
change-by-change basis. This had the negative side effect of allowing
some issues to be improperly labeled, based purely on which change was
last processed.

With this commit, all changes affecting an issue are examined while
determining which labels should be present on the issue, allowing for
more accurate labelling.